pyproject.toml
[project]
name = "promptoy"
version = "0.1.0"
description = ""
authors = [
    {name = "Yixuan-Wang", email = "tom-yx_wang@outlook.com"},
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "multimethod>=1.10",
]

[project.optional-dependencies]
openai = [
    "openai>=1.3",
]
anthropic = [
    "anthropic>=0.7.3",
]
dev = [
    "pytest>=7.4.3",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.1",
]
llama-cpp = [
    "llama-cpp-python>=0.2.20",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
